Remove use site sync management in favor of use sync sites context
